chore: move benchmarks to test package
