+++ tests/benchmark/otel/doc.go
