+++ tests/benchmark/otel/metrics_otel_generated.go
