+++ tests/benchmark/prometheus/metrics_prometheus_generated.go
